requirements.txt
Django==1.11
djangorestframework=3
Pillow==3